#000 | Fix issue with sync of dispatch receipt with cast exception

+++ goonj/src/main/java/org/avni_integration_service/goonj/repository/DispatchReceiptRepository.java
@@ -22,7 +22,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
@@ -108,12 +107,12 @@ public class DispatchReceiptRepository extends GoonjBaseRepository
         String itemName = typeOfMaterial.equals(CONTRIBUTED_ITEM)?
                 (String) entry.get(CONTRIBUTED_ITEM_NAME):
                 (typeOfMaterial.equals(KIT) ? (String) entry.get(KIT_NAME) : (String) entry.get(MATERIAL_NAME));
-        long dispatchedQuantity = entry.get(QUANTITY_DISPATCHED) != null ? (long) entry.get(QUANTITY_DISPATCHED) : 0L;
+        long dispatchedQuantity = entry.get(QUANTITY_DISPATCHED) != null ? (Integer) entry.get(QUANTITY_DISPATCHED) : 0L;
         boolean quantityMatching = entry.get(QUANTITY_MATCHING).equals("Yes");
         if (quantityMatching) {
             receivedQuantity = dispatchedQuantity;
         } else {
-            receivedQuantity = entry.get(QUANTITY) != null ? (long) entry.get(QUANTITY) : 0L;
+            receivedQuantity = entry.get(QUANTITY) != null ? (Integer) entry.get(QUANTITY) : 0L;
         }
         return new DispatchReceivedStatusLineItem(dispatchStatusLineItemId, mapTypeOfMaterial(entry), itemName,
                 dispatchStatusLineItemId, EMPTY_STRING, EMPTY_STRING, dispatchedQuantity, receivedQuantity);
